Conflicts:
	src/condor_schedd.V6/CMakeLists.txt
	src/condor_schedd.V6/schedd.cpp

+++ src/condor_utils/stl_string_utils.cpp
@@ -162,13 +162,17 @@ void trim( std::string &str )
 	if( str.empty() ) {
 		return;
 	}
-	int		begin = 0;
-	while ( begin < str.length() && isspace(str[begin]) ) { ++begin; }
+	unsigned	begin = 0;
+	while ( begin < str.length() && isspace(str[begin]) ) {
+		++begin;
+	}
 
-	int		end = str.length() - 1;
-	while ( end >= 0 && isspace(str[end]) ) { --end; }
+	int			end = str.length() - 1;
+	while ( end >= 0 && isspace(str[end]) ) {
+		--end;
+	}
 
-	if ( begin != 0 || end != str.length() - 1 ) {
+	if ( begin != 0 || end != (int)(str.length()) - 1 ) {
 		str = str.substr(begin, (end - begin) + 1);
 	}
 }
